refactor: move ClientControllerFactory to better folder and update imports
